Upgrade React to 19.1 and modernize all components
- Upgrade React from 0.13.3 to 19.1 with react-dom and prop-types packages
- Convert all React.createClass components to modern function components
- Rename .react.js files to .jsx for proper JSX recognition
- Switch to automatic JSX runtime and remove unnecessary React imports
- Simplify Vite configuration by removing JSX loader workarounds

This brings the codebase up to modern React standards with hooks support and improved performance.
