+++ src/entry.js
@@ -2,7 +2,7 @@
 import "./style.less";
 import React from "react";
 import FileActionCreators from "./actions/FileActionCreators";
-import SyntaxTreeComponent from "./views/SyntaxTree.react";
+import SyntaxTreeComponent from "./views/SyntaxTree.jsx";
 import Immutable from "immutable";
 
 const SyntaxTree = React.createFactory(SyntaxTreeComponent);
