+++ vite.config.js
@@ -4,16 +4,9 @@ import react from '@vitejs/plugin-react'
 export default defineConfig({
   plugins: [
     react({
-      jsxRuntime: 'classic',
-      include: '**/*.{jsx,tsx,js,ts}'
+      jsxRuntime: 'automatic'
     })
   ],
-  esbuild: {
-    jsx: 'transform',
-    loader: 'jsx',
-    include: /\.(js|jsx|ts|tsx)$/,
-    exclude: []
-  },
   css: {
     preprocessorOptions: {
       less: {
